Smart Borders
The idea of smart borders has been borrowed from sway.

When on a specific workspace (tag) there aren't at least two tiling
windows, borders aren't drawn.

Borders are very useful to easily understand which windows is in focus
and be able to distinguish one window from the other but when there is
only one window they become an ugly way to waste space.

If the floating layout is selected borders are always drawn.

Borders are always drawn for floating windows, even on a non floating
layout