Massive renderclients optimization
Calculate the number of clients only once each renderclients call, not
at only iteration of its main loop

Never count the clients if bw is going to be 0 anyway because the layout
is floating, monocle or borderpx == 0. Also those checks are done once
each renderclients.

Avoid separate if to check if borderpx == 0