Fix borders disappearing on milti mon
There's now a ybw for each monitor. This allows to calculate if borders
must be rendered or not on each monitor separately.

This fixes the bug of drawing borders depending on the number of clients
on the last arranged monitor (If a monitor with one client was rendered,
all other monitors would not have borders and vice versa)