nclients relative to mon just to make sure
nclients as global wasn't causing bugs because it wasn't being used
outside of tile() and arrange(), which are monitor relative operations.
However, if something was going to use it outside of those in the future
if would have had a wrong value, because it was being constantly
overwritten by calculations on different monitors

Also this eliminates the necessity for a global variable.