Major rework and optimization
These changes bring this patch to being way more similar to master while
also a lot more efficient

Drop Monitor.ybw and Monitor.nclients : all those calculations are done
in arrange(), tile() and monocle(). This eliminates the need to
calculate borders in rendermon() and therefore also the need to call
resize() in that function, which is quite heavy.