Render the focused fscrn client or everything else
Adapt to Monitor.focused

This follows the philosophy of the recent changes in the fullscreen
branch. Render only the fullscreen window when is focused, otherwise
keep rendering everything as usual. This allows more flexibility because
other clients can still be used even if one client is fullscreen, or
even multiple fullscreen client can exist in the same tag.